+++ scenedetect/__init__.py
@@ -41,7 +41,8 @@ from scenedetect.scene_manager import SceneManager, save_images
 from scenedetect.scene_detector import SceneDetector
 from scenedetect.frame_timecode import FrameTimecode
 from scenedetect.video_stream import VideoStream, VideoOpenFailure
-from scenedetect.backends import AVAILABLE_BACKENDS, VideoStreamCv2, VideoStreamAv
+from scenedetect.backends import (AVAILABLE_BACKENDS, VideoStreamCv2, VideoStreamAv,
+                                  VideoStreamMoviePy, VideoCaptureAdapter)
 from scenedetect.stats_manager import StatsManager, StatsFileCorrupt
 from scenedetect.detectors import ContentDetector, AdaptiveDetector, ThresholdDetector
 from scenedetect.video_splitter import split_video_ffmpeg, split_video_mkvmerge